chore: Update interdependent dependencies and revise command-line interface
  Updates click, Flask and several of its extensions, itsdangerous, Jinja2,
  MarkupSafe, SQLAlchemy, and Werkzeug.

  Removes the obsolete Flask-Script extension, as that has been superseded
  by the built-in Flask command-line interface `flask <COMMAND> ...`
  Reimplements `python manage.py test` as `flask test`.
  Makes progress on #43.
  Supersedes #63.
  Supersedes #90.
