+++ .gitignore
@@ -22,6 +22,7 @@ b000027.txt
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+api.log
 
 # Editor directories and files
 .idea
