+++ requirements-freeze.txt
@@ -1,24 +1,25 @@
 alembic==1.8.1
+blinker==1.6.3
 certifi==2023.7.22
 charset-normalizer==2.1.1
-click==7.1.2
+click==8.1.7
 cypari2==2.1.3
 cysignals==1.11.2
 Cython==0.29.36
-Flask==2.2.5
-flask-cors==3.0.10
-Flask-Executor==0.10.0
+Flask==3.0.0
+Flask-Cors==3.0.10
+Flask-Executor==1.0.0
 Flask-Login==0.5.0
 Flask-Migrate==2.7.0
-Flask-Script==2.0.6
-Flask-SQLAlchemy==2.5.1
+Flask-SQLAlchemy==3.1.1
 Flask-WTF==0.15.1
+greenlet==2.0.2
 idna==3.4
-itsdangerous==1.1.0
-Jinja2==2.11.3
+itsdangerous==2.1.2
+Jinja2==3.1.2
 jsonify==0.5
 Mako==1.2.2
-MarkupSafe==1.1.1
+MarkupSafe==2.1.3
 numpy==1.23.2
 psycopg2-binary==2.9.7
 python-dateutil==2.8.2
@@ -26,8 +27,9 @@ python-dotenv==0.20.0
 python-editor==1.0.4
 requests==2.28.1
 six==1.16.0
-SQLAlchemy==1.4.40
+SQLAlchemy==2.0.22
+typing_extensions==4.8.0
 urllib3==1.26.12
-Werkzeug==1.0.1
+Werkzeug==3.0.0
 wheel==0.41.2
 WTForms==2.3.3
